BackendDjango/requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1617745359837/work
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.7
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1617829946528/work
djangorestframework==3.12.4
gcloud==0.17.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
jws==0.1.3
oauth2client==3.0.0
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==2.0.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.15.0
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1602142927465/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
wincertstore==0.2
